# Fase-I/guidelines/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sistema de Análisis de Sentimiento de Reviews – Hotel OLH",
    "description": "API REST para el pipeline de NLP y el dashboard ejecutivo de análisis de sentimiento de reseñas hoteleras. Cubre los servicios: Data Verification & Cleaner, Sentiment Prediction, ABSAService (ABSA sobre tópicos hoteleros → `review_topicos`), Topic Identification (alternativa al ABSA para la misma etapa `topics_identified`), Metric Calculation (`POST /metricas/calcular` y `POST /metricas/recalcular` en MetricProcessor), Dashboard Backend (lecturas `GET /metricas/*`, `POST /metricas/recalcular-semestre` que orquesta recálculos vía MetricProcessor, `GET /metricas/archivos/entrada`, `POST /metricas/archivos/upload` y Reviews Explorer) y el listado filtrado de reviews.",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Tesis OLH",
      "email": "admin@olh.com"
    }
  },
  "servers": [
    {
      "url": "https://api.olh.com",
      "description": "Producción"
    },
    {
      "url": "http://localhost:8000",
      "description": "Desarrollo local (genérico)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Dashboard Backend (local): GET /api/v1/metricas/*, POST /api/v1/metricas/recalcular-semestre, GET /api/v1/metricas/archivos/entrada y POST /api/v1/metricas/archivos/upload"
    },
    {
      "url": "http://localhost:5004",
...
    },
    {
      "url": "http://localhost:5003",
...
    },
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/metricas/topicos": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Métricas por tópico del mes (Dashboard Backend)",
        "description": "Devuelve métricas por tópico leyendo `metricas_topico_mensual` y `topicos`.",
        "operationId": "getMetricasTopicosMensual",
        "parameters": [
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "$ref": "#/components/parameters/Anio"
          },
          {
            "$ref": "#/components/parameters/Mes"
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "clave",
                "adicional"
              ]
            },
            "description": "Filtrar por tipo de tópico. Vacío = todos."
          },
          {
            "name": "solo_alertas",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Si es true, retorna solo tópicos con alerta activa"
          }
        ],
        "responses": {
          "200": {
            "description": "Métricas de tópicos del período",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TopicoMetrica"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (hotel_id, anio, mes, tipo o solo_alertas)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "PARAMETROS_INVALIDOS",
                    "mensaje": "El parámetro 'mes' es requerido"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metricas/topicos/top5": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Top 5 tópicos más críticos (Dashboard Backend)",
        "description": "Devuelve los 5 tópicos con menor score del período leyendo `metricas_topico_mensual` y `topicos`.",
        "operationId": "getMetricasTopicosTop5Mensual",
        "parameters": [
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "$ref": "#/components/parameters/Anio"
          },
          {
            "$ref": "#/components/parameters/Mes"
          }
        ],
        "responses": {
          "200": {
            "description": "Top 5 tópicos más críticos del período",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "maxItems": 5,
                          "items": {
                            "$ref": "#/components/schemas/TopicoTop5"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (hotel_id, anio, mes)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "PARAMETROS_INVALIDOS",
                    "mensaje": "El parámetro 'mes' debe estar entre 1 y 12"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "HotelId": {
        "name": "hotel_id",
        "in": "query",
        "required": true,
        "schema": {
          "type": "integer"
        },
        "description": "ID del hotel"
      },
      "Anio": {
        "name": "anio",
        "in": "query",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 2020,
          "example": 2025
        },
        "description": "Año del período"
      },
      "Mes": {
        "name": "mes",
        "in": "query",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 12,
          "example": 11
        },
        "description": "Mes del período (1–12)"
      },
...
    },
    "responses": {
...
      "InternalError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EnvelopeError"
            },
            "example": {
              "ok": false,
              "data": null,
              "error": {
                "codigo": "ERROR_INTERNO",
                "mensaje": "Error inesperado en el procesamiento"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "type": "object",
            "nullable": true,
            "example": null
          }
        }
      },
      "EnvelopeError": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "error": {
            "type": "object",
            "required": [
              "codigo",
              "mensaje"
            ],
            "properties": {
              "codigo": {
                "type": "string",
                "example": "ARCHIVO_NO_ENCONTRADO"
              },
              "mensaje": {
                "type": "string",
                "example": "No se encontró el recurso solicitado"
              }
            }
          }
        }
      },
...
      "TopicoMetrica": {
        "type": "object",
        "properties": {
          "topico_id": {
            "type": "integer",
            "example": 7
          },
          "slug": {
            "type": "string",
            "example": "ruido"
          },
          "nombre": {
            "type": "string",
            "example": "Ruido"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "clave",
              "adicional"
            ],
            "example": "adicional"
          },
          "score_promedio": {
            "type": "number",
            "format": "float",
            "example": 52.3
          },
          "cambio_pct_vs_anterior": {
            "type": "number",
            "format": "float",
            "example": -5.2
          },
          "total_menciones": {
            "type": "integer",
            "example": 310
          },
          "alerta": {
            "type": "boolean",
            "example": true
          },
          "tendencia": {
            "type": "string",
            "enum": [
              "up",
              "down",
              "stable"
            ],
            "example": "down"
          }
        }
      },
      "TopicoTop5": {
        "type": "object",
        "properties": {
          "posicion": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 1
          },
          "slug": {
            "type": "string",
            "example": "ruido"
          },
          "nombre": {
            "type": "string",
            "example": "Ruido"
          },
          "score_promedio": {
            "type": "number",
            "format": "float",
            "example": 52.3
          },
          "alerta": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}